requirements.txt
Django>=5.0
django-tailwind>=3.8.0
Pillow>=10.1.0
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-filter>=23.5
channels>=4.0.0
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.0
django-htmx>=1.17.0
psycopg2-binary>=2.9.9
redis>=5.0.1
django-redis>=5.4.0
celery>=5.3.6
django-celery-beat>=2.5.0

# Production requirements
gunicorn>=21.2.0
mysqlclient>=2.2.1
whitenoise>=6.6.0
sentry-sdk>=1.39.1

# Render deployment requirements
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
gunicorn>=21.2.0
